rearrange some functions in modules
