+++ lambdas/api/index.js
@@ -5,42 +5,6 @@ const Api = require('sat-api-lib')
 const util = require('lambda-proxy-utils')
 const get = require('lodash.get')
 const es = require('../../lib/es')
-let esClient;
-
-function search(req, cb) {
-  const s = new Api(req, esClient);
-  //const encoding = get(req, 'headers.Accept-Encoding', null);
- 
-  s['search'](function (err, resp) {
-    if (err) {
-      console.log(err);
-      const res = new util.Response({ cors: true, statusCode: 400 });
-      return cb(null, res.send({ details: err.message }));
-    }
-    const res = new util.Response({ cors: true, statusCode: 200 });
-    return cb(null, res.send(resp));
-    /*
-    if (encoding && encoding.includes('gzip')) {
-      zlib.gzip(JSON.stringify(resp), function(error, gzipped) {
-        //if(error) context.fail(error);
-        const response = {
-          statusCode: 200,
-          body: gzipped.toString('base64'),
-          isBase64Encoded: true,
-          headers: {
-            'Content-Type': 'application/json',
-            'Content-Encoding': 'gzip'
-          }
-        };
-        return cb(error, response);
-      });
-    }
-    else {
-      return cb(null, res.send(resp));
-    }
-    */
-  });
-}
 
 
 module.exports.handler = function (event, context, cb) {
@@ -70,15 +34,20 @@ module.exports.handler = function (event, context, cb) {
   else if (method === 'GET' && event.queryStringParameters) {
     payload.query = event.queryStringParameters;
   }
+    
+  es.client().then((esClient) => {
 
-  if (!esClient) {
-    console.log('connecting to ES');
-    es.connect().then((client) => {
-      esClient = client;
-      search(payload, cb);
-    });
-  }
-  else {
-    search(payload, cb);
-  }
-};
+    const s = new Api.search(payload, esClient);
+    //const encoding = get(req, 'headers.Accept-Encoding', null);
+    
+    s['search'](function (err, resp) {
+      if (err) {
+        console.log(err);
+        const res = new util.Response({ cors: true, statusCode: 400 });
+        return cb(null, res.send({ details: err.message }));
+      }
+      const res = new util.Response({ cors: true, statusCode: 200 });
+      return cb(null, res.send(resp));
+    }); 
+  })
+}
