+++ lib/es.js
@@ -3,9 +3,32 @@
 const AWS = require('aws-sdk')
 const httpAwsEs = require('http-aws-es')
 const elasticsearch = require('elasticsearch')
+const through2 = require('through2')
+const ElasticsearchWritableStream = require('elasticsearch-writable-stream')
+const pump = require('pump')
+
+const index = process.env.ES_INDEX || 'items'
+
+let esClient;
+
+
+async function Client() {
+  if (!esClient) {
+    console.log('connecting to elasticsearch');
+    connect().then((client) => {
+      esClient = client
+      return esClient
+    })
+  } else {
+    console.log('using existing elasticsearch connection')
+    return esClient
+  }
+}
+
 
 // Connect to an Elasticsearch cluster
 async function connect() {
+  
   let esConfig
   let client
 
@@ -134,9 +157,56 @@ async function deleteIndex(esClient, index) {
   })
 }*/
 
+function streamToEs(esClient, stream, transform, type='sat', id='id') {
+  // Given an input stream and a transform, write records to an elasticsearch instance
+
+  var n_records = 0
+  var n_csv = 0
+  var n_transformed = 0
+
+  var toEs = through2({'objectMode': true, 'consume': true}, function(data, encoding, next) {
+    var record = {
+      index: index,
+      type: type, 
+      id: data['id'],
+      action: 'update',
+      _retry_on_conflict: 3,
+      body: {
+        doc: data,
+        doc_as_upsert: true
+      }
+    }
+    this.push(record)
+    next()
+  })
+
+  var esStream = new ElasticsearchWritableStream(esClient, {
+    highWaterMark: 100,
+    flushTimeout: 1000
+  })
+
+  return new Promise((resolve, reject) => {
+    pump(stream, transform, toEs, esStream, function(err) {
+        console.log('stream done', err)
+          if (err) {
+            console.log('error:', err)
+            reject(n_transformed)
+          } else {
+            console.log(`Finished: ${n_records} csv records, ${n_transformed} transformed, `)
+            resolve(n_transformed)
+          }
+        })
+
+    // count records
+    stream.on('data', (data) => {n_records++})
+    toEs.on('data', (data) => {n_transformed++})
+  })
+}
+
+module.exports.client = Client
 module.exports.reindex = reindex
-module.exports.connect = connect
 module.exports.listIndices = listIndices
 module.exports.putMapping = putMapping
 module.exports.deleteIndex = deleteIndex
+module.exports.streamToEs = streamToEs
 //module.exports.saveRecords = saveRecords
