+++ lib/metadata.js
@@ -7,9 +7,7 @@ const AWS = require('aws-sdk')
 const queue = require('async.queue')
 const es = require('./es')
 const zlib = require('zlib')
-const through2 = require('through2')
-const ElasticsearchWritableStream = require('elasticsearch-writable-stream')
-const pump = require('pump')
+
 let esClient
 const got = require('got')
 const stream = require('stream')
@@ -19,53 +17,6 @@ const s3 = new AWS.S3()
 const index = process.env.ES_INDEX || 'items'
 
 
-function streamToEs(stream, transform, type='sat', id='id') {
-  // Given an input stream and a transform, write records to an elasticsearch instance
-
-  var n_records = 0
-  var n_csv = 0
-  var n_transformed = 0
-
-  var toEs = through2({'objectMode': true, 'consume': true}, function(data, encoding, next) {
-    var record = {
-      index: index,
-      type: type, 
-      id: data['id'],
-      action: 'update',
-      _retry_on_conflict: 3,
-      body: {
-        doc: data,
-        doc_as_upsert: true
-      }
-    }
-    this.push(record)
-    next()
-  })
-
-  var esStream = new ElasticsearchWritableStream(esClient, {
-    highWaterMark: 100,
-    flushTimeout: 1000
-  })
-
-  return new Promise((resolve, reject) => {
-    pump(stream, transform, toEs, esStream, function(err) {
-        console.log('stream done', err)
-          if (err) {
-            console.log('error:', err)
-            reject(n_transformed)
-          } else {
-            console.log(`Finished: ${n_records} csv records, ${n_transformed} transformed, `)
-            resolve(n_transformed)
-          }
-        })
-
-    // count records
-    stream.on('data', (data) => {n_records++})
-    toEs.on('data', (data) => {n_transformed++})
-  })
-}
-
-
 function processCsvFile(bucket, key, transform) {
   // get the csv file s3://${bucket}/${key}
  
@@ -73,80 +24,7 @@ function processCsvFile(bucket, key, transform) {
   const csvStream = csv.parse({ headers: true, objectMode: true })
   s3.getObject({Bucket: bucket, Key: key}).createReadStream().pipe(csvStream)
 
-  return streamToEs(csvStream, transform)
-}
-
-
-// kick off processing next CSV file
-function invokeLambda(bucket, key, satellite, nextFileNum, lastFileNum, arn, retries) {
-    // figure out if there's a next file to process
-    
-    if (nextFileNum && arn) {
-      const stepfunctions = new AWS.StepFunctions()
-      const params = {
-        stateMachineArn: arn,
-        input: JSON.stringify({ bucket, key, satellite, currentFileNum: nextFileNum, lastFileNum, arn, retries}),
-        name: `csv_${satellite}_${nextFileNum}_${Date.now()}`
-      }
-      stepfunctions.startExecution(params, function(err, data) {
-        if (err) {
-          console.log(err, err.stack)
-        } else {
-          console.log(`launched ${JSON.stringify(params)}`)
-        }
-      }) 
-    }
-}
-
-
-// Process 1 or more CSV files by processing one at a time, then invoking the next
-function processCsvFiles(bucket, key, satellite, transform, cb, currentFileNum=0, lastFileNum=0, arn=null, retries=0) {
-  const maxRetries = 10
-
-  const currentKey = `${key}/${satellite}_${currentFileNum}.csv`;
-  var nextFileNum = (currentFileNum < lastFileNum) ? currentFileNum + 1 : null
-  //invokeLambda(bucket, key, satellite, currentFileNum, lastFileNum, arn)
-
-  processCsvFile(
-    bucket, currentKey, transform
-  ).then((n_scenes) => {
-    invokeLambda(bucket, key, satellite, nextFileNum, lastFileNum, arn, 0)
-    cb()
-  }).catch((e) => {
-    // if CSV failed, try it again
-    if (retries < maxRetries) {
-      invokeLambda(bucket, key, satellite, currentFileNum, lastFileNum, arn, retries + 1)
-    } else {
-      // log and move onto the next one
-      console.log(`error: maxRetries hit in file ${currentFileNum}`)
-      invokeLambda(bucket, key, satellite, nextFileNum, lastFileNum, arn, 0)
-    }
-    cb()
-  })
-}
-
-
-function invokeLambda2(satellite, firstFileNum, lastFileNum, arn, cb) {
-  const stepfunctions = new AWS.StepFunctions()
-  const prefix = process.env.prefix || 'sat-api-dev'
-  if (firstFileNum > lastFileNum) return
-  const inputs = {
-    bucket: process.env.bucket || 'sat-api',
-    key: `${prefix}/csv/${satellite}`,
-    satellite,
-    currentFileNum: firstFileNum,
-    lastFileNum,
-    arn
-  }
-  const params = {
-    stateMachineArn: arn,
-    input: JSON.stringify(inputs),
-    name: `csv_${satellite}_${firstFileNum}_${Date.now()}`
-  }
-  stepfunctions.startExecution(params, function(err, data) {
-    console.log(`${params.name} step function launched: ${JSON.stringify(inputs)}`);
-    if (err) console.log(`step function launch error: ${err}`)
-  })
+  return es.streamToEs(esClient, csvStream, transform)
 }
 
 
@@ -266,6 +144,80 @@ function split(satellite, arn, maxFiles, linesPerFile, maxLambdas, cb) {
 }
 
 
+
+
+// kick off processing next CSV file
+function invokeLambda(bucket, key, satellite, nextFileNum, lastFileNum, arn, retries) {
+    // figure out if there's a next file to process
+    
+    if (nextFileNum && arn) {
+      const stepfunctions = new AWS.StepFunctions()
+      const params = {
+        stateMachineArn: arn,
+        input: JSON.stringify({ bucket, key, satellite, currentFileNum: nextFileNum, lastFileNum, arn, retries}),
+        name: `csv_${satellite}_${nextFileNum}_${Date.now()}`
+      }
+      stepfunctions.startExecution(params, function(err, data) {
+        if (err) {
+          console.log(err, err.stack)
+        } else {
+          console.log(`launched ${JSON.stringify(params)}`)
+        }
+      }) 
+    }
+}
+
+
+// Process 1 or more CSV files by processing one at a time, then invoking the next
+function processCsvFiles(bucket, key, satellite, transform, cb, currentFileNum=0, lastFileNum=0, arn=null, retries=0) {
+  const maxRetries = 10
+
+  const currentKey = `${key}/${satellite}_${currentFileNum}.csv`;
+  var nextFileNum = (currentFileNum < lastFileNum) ? currentFileNum + 1 : null
+  //invokeLambda(bucket, key, satellite, currentFileNum, lastFileNum, arn)
+
+  processCsvFile(
+    bucket, currentKey, transform
+  ).then((n_scenes) => {
+    invokeLambda(bucket, key, satellite, nextFileNum, lastFileNum, arn, 0)
+    cb()
+  }).catch((e) => {
+    // if CSV failed, try it again
+    if (retries < maxRetries) {
+      invokeLambda(bucket, key, satellite, currentFileNum, lastFileNum, arn, retries + 1)
+    } else {
+      // log and move onto the next one
+      console.log(`error: maxRetries hit in file ${currentFileNum}`)
+      invokeLambda(bucket, key, satellite, nextFileNum, lastFileNum, arn, 0)
+    }
+    cb()
+  })
+}
+
+function invokeLambda2(satellite, firstFileNum, lastFileNum, arn, cb) {
+  const stepfunctions = new AWS.StepFunctions()
+  const prefix = process.env.prefix || 'sat-api-dev'
+  if (firstFileNum > lastFileNum) return
+  const inputs = {
+    bucket: process.env.bucket || 'sat-api',
+    key: `${prefix}/csv/${satellite}`,
+    satellite,
+    currentFileNum: firstFileNum,
+    lastFileNum,
+    arn
+  }
+  const params = {
+    stateMachineArn: arn,
+    input: JSON.stringify(inputs),
+    name: `csv_${satellite}_${firstFileNum}_${Date.now()}`
+  }
+  stepfunctions.startExecution(params, function(err, data) {
+    console.log(`${params.name} step function launched: ${JSON.stringify(inputs)}`);
+    if (err) console.log(`step function launch error: ${err}`)
+  })
+}
+
+
 function update(event, transform, cb) {
   const bucket = _.get(event, 'bucket')
   const key = _.get(event, 'key')
@@ -275,18 +227,12 @@ function update(event, transform, cb) {
   const arn = _.get(event, 'arn', null)
   const retries = _.get(event, 'retries', 0)
 
-  if (!esClient) {
-    es.connect().then((client) => {
-      esClient = client
-      es.putMapping(esClient, index).catch((err) => {})
-      processCsvFiles(bucket, key, sat, transform, cb, currentFileNum, lastFileNum, arn, retries)
-    }).catch(err => {
-      console.log(err)
-    })
-  }
-  else {
+  es.connect().then((esClient) => {
+    es.putMapping(esClient, index).catch((err) => {})
     processCsvFiles(bucket, key, sat, transform, cb, currentFileNum, lastFileNum, arn, retries)
-  }
+  }).catch(err => {
+    console.log(err)
+  })
 }
